django/requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
contextlib2==0.5.5
cryptography==2.0.3
Django==1.11.3
django-admin-row-actions==0.0.2
django-analytical==2.2.2
django-annoying==0.10.3
django-anymail==0.11.1
django-attachments==1.1.1
django-compat==1.0.14
django-dbbackup==3.1.3
django-extensions==1.8.1
django-fancy-cache==0.9.0
django-formtools==2.0
django-geojson==2.10.0
django-hijack==2.1.4
django-leaflet==0.22.0
django-localflavor==1.5.2
django-material==1.0.0
django-model-utils==3.0.0
django-taggit==0.22.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Fabric3==1.13.1.post1
html2text==2016.9.19
idna==2.6
jdcal==1.3
jsonfield==2.0.2
lml==0.0.1
nameparser==0.5.3
olefile==0.44
openpyxl==2.4.8
paramiko==2.3.0
phonenumbers==8.6.0
Pillow==4.2.1
psycopg2==2.7.1
pyasn1==0.3.5
pycparser==2.18
pyexcel-io==0.5.1
pyexcel-xlsx==0.4.1
PyJWT==1.5.3
PyNaCl==1.1.2
pytz==2017.2
raven==6.1.0
requests==2.18.4
six==1.11.0
urllib3==1.22
